mvp/agents/requirements.txt
agent-framework==1.0.0rc4
agent-framework-azure-ai==1.0.0rc4
agent-framework-orchestrations==1.0.0b260311
openai>=2.28.0
pydantic>=2.0
pydantic-settings>=2.10.1
python-dotenv>=1.0
azure-identity>=1.25.0
httpx>=0.27.0
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
msal>=1.31.1
PyJWT[crypto]>=2.10.1
